Thin remaining small endpoints; coins module; harden import ratchet
- services for bot, users, coin_deposits, rss
- award/deduct coins live in app/coins.py (crud-safe)
- import ratchet: api→crud/responders is ERROR except login/people allowlist
- 0 hard errors
