+++ chafan_core/app/api/api_v1/endpoints/bot.py
@@ -2,13 +2,11 @@ from typing import Any
 
 from fastapi import APIRouter, Depends
 
-from chafan_core.app import crud, schemas, security
+from chafan_core.app import schemas
 from chafan_core.app.api import deps
-from chafan_core.app.config import settings
 from chafan_core.app.infra.request_context import RequestContext
 from chafan_core.app.schemas.security import VerifiedTelegramID, VerifyTelegramID
-from chafan_core.app.services import submissions as submissions_service
-from chafan_core.utils.base import HTTPException_
+from chafan_core.app.services import bot as bot_service
 
 router = APIRouter()
 
@@ -19,27 +17,7 @@ def verify_telegram_id(
     ctx: RequestContext = Depends(deps.get_request_context),
     data: VerifyTelegramID,
 ) -> Any:
-    if (
-        (not settings.OFFICIAL_BOT_SECRET)
-        or data.verifier_secret != settings.OFFICIAL_BOT_SECRET
-        or (not security.check_digit_verification_code(data.email, data.code))
-    ):
-        raise HTTPException_(
-            status_code=400,
-            detail="Unauthenticated.",
-        )
-    user = crud.user.get_by_email(ctx.get_db(), email=data.email)
-    if user is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="Invalid email.",
-        )
-    crud.user.update(
-        ctx.get_db(),
-        db_obj=user,
-        obj_in={"verified_telegram_user_id": data.telegram_id},
-    )
-    return schemas.VerifyTelegramResponse(handle=user.handle)
+    return bot_service.verify_telegram_id(ctx, data)
 
 
 @router.post("/submissions/", response_model=schemas.Submission)
@@ -49,22 +27,6 @@ def create_submission(
     submission_in: schemas.SubmissionCreate,
     verified_id: VerifiedTelegramID,
 ) -> Any:
-    if (
-        not settings.OFFICIAL_BOT_SECRET
-    ) or verified_id.verifier_secret != settings.OFFICIAL_BOT_SECRET:
-        raise HTTPException_(
-            status_code=400,
-            detail="Unauthenticated.",
-        )
-    user = crud.user.get_by_telegram_id(
-        ctx.get_db(), telegram_id=verified_id.telegram_id
+    return bot_service.create_submission_as_bot(
+        ctx, submission_in=submission_in, verified_id=verified_id
     )
-    if not user:
-        raise HTTPException_(
-            status_code=400,
-            detail="User doesn't exist.",
-        )
-    _submission, data = submissions_service.create_submission(
-        ctx, submission_in=submission_in, author=user
-    )
-    return data
