+++ chafan_core/app/crud/crud_coin_payment.py
@@ -4,8 +4,7 @@ from typing import List, Optional
 from fastapi.encoders import jsonable_encoder
 from sqlalchemy.orm import Session
 
-from chafan_core.app import models
-from chafan_core.app.services import reputation as rep_manager
+from chafan_core.app import coins, models
 from chafan_core.app.crud.base import CRUDBase
 from chafan_core.app.models.coin_payment import CoinPayment
 from chafan_core.app.schemas.coin_payment import CoinPaymentCreate, CoinPaymentUpdate
@@ -20,8 +19,8 @@ class CRUDCoinPayment(CRUDBase[CoinPayment, CoinPaymentCreate, CoinPaymentUpdate
         payer: models.User,
         payee: models.User
     ) -> CoinPayment:
-        rep_manager.deduct_coins(db, payer, obj_in.amount, "coin_payment_out")
-        rep_manager.award_coins(db, payee, obj_in.amount, "coin_payment_in")
+        coins.deduct_coins(db, payer, obj_in.amount, "coin_payment_out")
+        coins.award_coins(db, payee, obj_in.amount, "coin_payment_in")
         payment = CoinPayment(
             **jsonable_encoder(obj_in),
             created_at=datetime.datetime.now(tz=datetime.timezone.utc),
