+++ chafan_core/app/services/__init__.py
@@ -7,7 +7,9 @@ from chafan_core.app.services import (
     article_columns,
     articles,
     audit,
+    bot,
     channels,
+    coin_deposits,
     coin_payments,
     comments,
     discovery,
@@ -27,11 +29,13 @@ from chafan_core.app.services import (
     reports,
     reputation,
     rewards,
+    rss,
     search,
     sites,
     submission_suggestions,
     submissions,
     topics,
+    users,
     viewcounts,
     webhook_delivery,
     webhooks,
@@ -44,7 +48,9 @@ __all__ = [
     "article_columns",
     "articles",
     "audit",
+    "bot",
     "channels",
+    "coin_deposits",
     "coin_payments",
     "comments",
     "discovery",
@@ -64,11 +70,13 @@ __all__ = [
     "reports",
     "reputation",
     "rewards",
+    "rss",
     "search",
     "sites",
     "submission_suggestions",
     "submissions",
     "topics",
+    "users",
     "viewcounts",
     "webhook_delivery",
     "webhooks",
